+++ tests/v4/verify.ts
@@ -9,8 +9,6 @@ import { sign } from '../../src/v4/sign';
 import { test } from 'uvu';
 import { verify } from '../../src/v4/verify';
 
-globalThis.crypto = crypto;
-
 const keys = {
     secretKey: "k4.secret.FgbULh0ylLoBsG6KRi2ZM0ZDzNMgaCBp1jB0sbf8OXGBf_1Cd0wyDa76n-iN0qGj0vaYSu5QXdZhbj5lUWhkyA",
     publicKey: "k4.public.gX_9QndMMg2u-p_ojdKho9L2mEruUF3WYW4-ZVFoZMg",